refactor: user-channel.ts, list-channel.ts を CQRS パターンで分割
user-channel.ts (455行) と list-channel.ts (352行) を queries/commands に分割し、
元ファイルを re-export のみに変更。全 import パスは後方互換を維持。

Closes #190
